feat(icon): implement Icon component and remove unused styles
