+++ src/components/footer/footer.module.css
@@ -24,5 +24,5 @@
 }
 
 .link {
-
+  color: var(--color-black);
 }
