--- src/components/icon/icon.module.css
@@ -1,3 +1,0 @@
-.icon {
-  color: var(--color-black);
-}
